requirements.txt
requests>=2.31.0
pandas>=2.2.0
matplotlib>=3.9.0
Pillow>=10.0.0

